+++ examples/sveltekit/src/routes/login/+page.ts
@@ -4,5 +4,5 @@ import type { PageLoad } from './$types';
 
 export const load: PageLoad = async (event) => {
 	const user = await getUser(event);
-	if (user) throw redirect(302, '/profile');
+	if (user) throw redirect(302, '/');
 };