+++ examples/sveltekit/vite.config.js
@@ -2,10 +2,7 @@ import { sveltekit } from '@sveltejs/kit/vite';
 
 /** @type {import('vite').UserConfig} */
 const config = {
-	plugins: [sveltekit()],
-	ssr: {
-		noExternal: ["@lucia-auth/sveltekit"]
-	}
+	plugins: [sveltekit()]
 };
 
 export default config;