+++ packages/astro/src/ambient.d.ts
@@ -1,0 +1,11 @@
+/// <reference types="lucia-auth" />
+declare namespace Lucia {
+	export type UserAttributes = {
+		username: string
+	}
+	export type Auth = {
+		configs: {
+			transformUserData: () => {}
+		}
+	};
+}