+++ packages/express/src/index.ts
@@ -4,22 +4,15 @@ import { convertExpressRequestToStandardRequest } from "./request.js";
 
 export const handleMiddleware = (auth: Auth) => {
 	const middleware = async (req: Request, res: Response, next: NextFunction) => {
-		req.app.locals.setSession = (session: Session) => {
-			const serializedCookies = auth.createSessionCookies(session);
-			res.append("Set-Cookie", serializedCookies.join());
-		};
-		req.app.locals.clearSession = () => {
-			const serializedCookies = auth.createBlankSessionCookies();
-			res.append("Set-Cookie", serializedCookies.join());
-		};
-		const setCookie = (stringifiedCookie: string) => {
-			res.append("Set-Cookie", stringifiedCookie);
+		req.app.locals.setSession = (session: Session | null) => {
+			const cookies = auth.createSessionCookies(session);
+			res.append("Set-Cookie", cookies.map((val) => val.serialize()).toString());
 		};
 		req.app.locals.getSession = async () => {
 			try {
 				return await auth.validateRequest(
 					convertExpressRequestToStandardRequest(req, auth),
-					setCookie
+					req.app.locals.setSession as (session: Session | null) => void
 				);
 			} catch {
 				return null;
@@ -35,7 +28,7 @@ export const handleMiddleware = (auth: Auth) => {
 			try {
 				return await auth.getSessionUserFromRequest(
 					convertExpressRequestToStandardRequest(req, auth),
-					setCookie
+					req.app.locals.setSession as (session: Session | null) => void
 				);
 			} catch {
 				return {