+++ packages/lucia-auth/src/auth/cookie.ts
@@ -1,35 +1,30 @@
 import { Auth, Session } from "../types.js";
-import { serialize } from "../utils/cookie.js";
+import { Cookie } from "../utils/cookie.js";
 
-type CreateSessionCookies = (session: Session) => string[];
+type CreateSessionCookies = (session: Session | null) => Cookie[];
 
 export const createSessionCookiesFunction = (auth: Auth) => {
 	const createSessionCookies: CreateSessionCookies = (session) => {
-		return auth.configs.sessionCookieOptions.map((option) => {
-			return serialize("auth_session", session.sessionId, {
-				...option,
-				httpOnly: true,
-				expires: new Date(session.idlePeriodExpires),
-				secure: auth.configs.env === "PROD"
+		if (session) {
+			return auth.configs.sessionCookieOptions.map((option) => {
+				return new Cookie("auth_session", session.sessionId, {
+					...option,
+					httpOnly: true,
+					expires: new Date(session.idlePeriodExpires),
+					secure: auth.configs.env === "PROD"
+				});
 			});
-		});
+		}
+		return [...auth.configs.sessionCookieOptions, ...auth.configs.deleteCookieOptions].map(
+			(option) => {
+				return new Cookie("auth_session", "", {
+					...option,
+					httpOnly: true,
+					maxAge: 0,
+					secure: auth.configs.env === "PROD"
+				});
+			}
+		);
 	};
 	return createSessionCookies;
-};
-
-type CreateBlankSessionCookies = () => string[];
-
-export const createBlankSessionCookiesFunction = (auth: Auth) => {
-	const createBlankSessionCookies: CreateBlankSessionCookies = () => {
-		const options = [...auth.configs.sessionCookieOptions, ...auth.configs.deleteCookieOptions];
-		return options.map((option) => {
-			return serialize("auth_session", "", {
-				...option,
-				httpOnly: true,
-				maxAge: 0,
-				secure: auth.configs.env === "PROD"
-			});
-		});
-	};
-	return createBlankSessionCookies;
 };