+++ packages/sveltekit/src/server/endpoints/index.ts
@@ -1,15 +1,1 @@
-export class ErrorResponse extends Response {
-	constructor(error: Error) {
-		console.error(error);
-		super(
-			JSON.stringify({
-				message: error.message
-			}),
-			{
-				status: 500
-			}
-		);
-	}
-}
-
 export { handleLogoutRequest } from "./logout.js";