+++ test-apps/nextjs/pages/_app.tsx
@@ -3,5 +3,10 @@ import "../styles/base.css";
 import type { AppProps } from "next/app";
 
 export default function App({ Component, pageProps }: AppProps) {
-	return <Component {...pageProps} />;
+	return (
+		<>
+			<h1>Lucia + Next.js demo</h1>
+			<Component {...pageProps} />
+		</>
+	);
 }