+++ test-apps/sveltekit/src/routes/+page.ts
@@ -1,8 +1,8 @@
 import { redirect } from '@sveltejs/kit';
 import { getUser } from '@lucia-auth/sveltekit/load';
-import type { PageLoad } from './$types';
+import type { PageLoad } from './$types.js';
 
 export const load: PageLoad = async (event) => {
 	const user = await getUser(event);
-	if (user) throw redirect(302, '/profile');
+	if (!user) throw redirect(302, '/login');
 };